+++ src/Symfony/Component/Security/Http/Authenticator/Passport/SelfValidatingPassport.php
@@ -26,21 +26,11 @@ use Symfony\Component\Security\Http\Authenticator\Passport\Badge\UserBadge;
 class SelfValidatingPassport extends Passport
 {
     /**
-     * @param UserBadge        $userBadge
      * @param BadgeInterface[] $badges
      */
-    public function __construct($userBadge, array $badges = [])
+    public function __construct(UserBadge $userBadge, array $badges = [])
     {
-        if ($userBadge instanceof UserInterface) {
-            trigger_deprecation('symfony/security-http', '5.2', 'The 1st argument of "%s" must be an instance of "%s", support for "%s" will be removed in symfony/security-http 5.3.', __CLASS__, UserBadge::class, UserInterface::class);
-
-            $this->user = $userBadge;
-        } elseif ($userBadge instanceof UserBadge) {
-            $this->addBadge($userBadge);
-        } else {
-            throw new \TypeError(sprintf('Argument 1 of "%s" must be an instance of "%s", "%s" given.', __METHOD__, UserBadge::class, get_debug_type($userBadge)));
-        }
-
+        $this->addBadge($userBadge);
         foreach ($badges as $badge) {
             $this->addBadge($badge);
         }